Port entrypoint changes from 2.x to 3.0
