+++ images/airflow/3.0.2/python/mwaa/logging/config.py
@@ -32,10 +32,17 @@ from mwaa.logging import cloudwatch_handlers
 from mwaa.logging.cloudwatch_handlers import CloudWatchRemoteTaskLogger
 from mwaa.utils import qualified_name
 
+CONSOLE_LOG_LEVEL = os.environ.get('AIRFLOW_CONSOLE_LOG_LEVEL', 'INFO')
+
 # We adopt the default logging configuration from Airflow and do the necessary changes
 # to setup logging with CloudWatch Logs.
 LOGGING_CONFIG = {
     **DEFAULT_LOGGING_CONFIG,
+    'root': {
+        'handlers': ['console'],
+        'level': CONSOLE_LOG_LEVEL,
+        'filters': ['mask_secrets'],
+    }
 }
 
 DAG_PROCESSOR_MANAGER_LOG_LOCATION = "/usr/local/airflow/logs/processor_manager/dag-processor-manager.log"
@@ -196,7 +203,7 @@ def _configure():
             comp,
             log_group_arn=log_group_arn,
             log_stream_name_prefix=comp.lower(),
-            log_level=log_level,
+            log_level="DEBUG",  # Customer Log Level handled at root logger,
             logging_enabled=logging_enabled,
         )
         _configure_subprocesses_logging(
